applications/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
azure-cosmos==4.7.0
python-dotenv==1.0.1
